% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_functions.R
\name{calculate_cov_matrix}
\alias{calculate_cov_matrix}
\title{Beregn årlig kovariansmatrix ud fra daglige log-afkast.}
\usage{
calculate_cov_matrix(data_list, threshold = 0.001)
}
\arguments{
\item{data_list}{En liste af xts-objekter.}

\item{threshold}{Numerisk grænse for minimumsvarians (default 1e-2).}
}
\value{
Den filtrerede annualiserede kovariansmatrix.
}
\description{
Filtrer aktiver med varians under en vis grænse (threshold).
}
