% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_functions.R
\name{calculate_expected_return}
\alias{calculate_expected_return}
\title{Beregn forventet årligt afkast ud fra daglige log-afkast.}
\usage{
calculate_expected_return(data)
}
\arguments{
\item{data}{En liste af xts-objekter.}
}
\value{
En navngivet liste med forventede årlige afkast.
}
\description{
Beregn forventet årligt afkast ud fra daglige log-afkast.
}
