% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_functions.R
\name{load_data}
\alias{load_data}
\title{Load data from Yahoo Finance}
\usage{
load_data(s, ticker_list)
}
\arguments{
\item{s}{Startdato (fx "2024-01-01").}

\item{ticker_list}{En karaktervektor af ticker-strenge.}
}
\value{
En liste med xts-objekter indeholdende prisdata.
}
\description{
Henter kursdata for en række tickers fra Yahoo.
}
