% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_functions.R
\docType{data}
\name{ticker_list}
\alias{ticker_list}
\title{Standard tickers for porteføljen}
\format{
An object of class \code{character} of length 34.
}
\usage{
ticker_list
}
\value{
En karaktervektor med ticker-strenge.
}
\description{
^OMXC25 er IKKE med her, da vi kun vil bruge den som benchmark (ikke i selve optimeringen).
}
\keyword{datasets}
